New component Editor (wrapping react-annotation-tool)
