--- src/lib/TextInput.css
@@ -1,14 +1,0 @@
-.simple-form-group {
-  margin-bottom: 1rem;
-}
-.simple-text-label {
-  display: block;
-  color: red;
-}
-.simple-text-input {
-  display: inline-block;
-  margin-bottom: 0.5rem;
-  font-size: 16px;
-  font-weight: 400;
-  color: rgb(33, 37, 41);
-}
