--- src/lib/TextInput.js
@@ -1,16 +1,0 @@
-import React from "react";
-import "./TextInput.css";
-
-const TextInput = ({ type = "text", label, value, onChange }) => (
-  <div className="simple-form-group">
-    {label && <label className="simple-text-label">{label}</label>}
-    <input
-      type={type}
-      className="simple-text-input"
-      value={value}
-      onChange={e => onChange && onChange(e.target.value)}
-    />
-  </div>
-);
-
-export default TextInput;
