+++ src/lib/index.js
@@ -1,3 +1,3 @@
-import TextInput from "./TextInput";
+import Editor from "./Editor";
 import { TextAnnotator } from 'react-text-annotate';
-export { TextInput, TextAnnotator };
+export { TextAnnotator, Editor };
